updates cluster node status at the end
