+++ _common/micro/MicroService.js
@@ -29,7 +29,8 @@ MicroService.prototype.init = function () {
       this.getSystemCodes.bind(this),
       this.establishQConnection.bind(this),
       this.connectExchange.bind(this),
-      this.connectToQueue.bind(this)
+      this.connectToQueue.bind(this),
+      this.updateClusterNodeStatus.bind(this)
     ],
     function (err) {
       if (err)
@@ -371,3 +372,34 @@ function _rejectMessage(bag, next) {
     bag.ackWaitTimeMS
   );
 }
+
+MicroService.prototype.updateClusterNodeStatus = function (next) {
+  if (!config.nodeId) {
+    logger.verbose(util.format('%s| Skipping cluster node status update ' +
+      'as no nodeId is present', msName));
+    return next();
+  }
+  logger.verbose(util.format('%s| updating cluster node status', msName));
+
+  var update = {
+    statusCode: global.systemCodesByName.success.code,
+    execImage: config.execImage,
+    stepId: null
+  };
+
+  this.publicAdapter.putClusterNodeById(config.nodeId,
+    update,
+    function (err) {
+      if (err) {
+        logger.warn(
+          util.format('Failed to update status of cluster node %s ' +
+            'with err %s', config.nodeId, err)
+        );
+        return next(true);
+      }
+      logger.verbose(util.format('%s| updated cluster node status to success',
+        msName));
+      return next();
+    }
+  );
+};
