+++ _common/micro/checkHealth.js
@@ -5,7 +5,6 @@ module.exports = self;
 var checkAMQP = require('./healthChecks/checkAMQP.js');
 var checkShippableApi = require('./healthChecks/checkShippableApi.js');
 var validateNode = require('./healthChecks/validateNode.js');
-var updateNodeStatus = require('./healthChecks/updateNodeStatus.js');
 
 function checkHealth(callback) {
   var bag = {};
@@ -20,7 +19,6 @@ function checkHealth(callback) {
   async.series([
       checkAMQP.bind(null, params),
       checkShippableApi.bind(null, params),
-      updateNodeStatus.bind(null, params),
       validateNode.bind(null, params)
     ],
     function (err) {
